+++ app/policies/third_party_project_policy.rb
@@ -1,0 +1,2 @@
+class ThirdPartyProjectPolicy < ProjectPolicy
+end