+++ CHANGELOG.md
@@ -113,6 +113,7 @@
 - XML file for projects now shows the identifiers for the parent activities.
 - Fix descriptive labels on action links
 - Remove `reference` from Transactions
+- Add level D activities (third-party projects)
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-4...HEAD
 [release-4]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-3...release-4